+++ pkg/deployment/resources/inspector/am_v1.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2022 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -79,6 +79,14 @@ func (p *arangoMembersInspectorV1) GetSimple(name string) (*api.ArangoMember, bo
 	}
 
 	return arangoMember, true
+}
+
+func (p *arangoMembersInspectorV1) GetSimpleOptional(name string) *api.ArangoMember {
+	a, ok := p.GetSimple(name)
+	if !ok {
+		return &api.ArangoMember{}
+	}
+	return a
 }
 
 func (p *arangoMembersInspectorV1) Iterate(action ins.Action, filters ...ins.Filter) error {