+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - (Feature) Add basic metrics for ArangoDeploymentReplication CR
 - (Bugfix) Use ArangoMember as owner reference for syncmaster secrets instead of Deployment
 - (Improvement) Remove PodSchedulingFailure condition instead of setting to false, restart pod if it could not be scheduled
+- (Feature) Add ArangoMember overrides
 
 ## [1.2.32](https://github.com/arangodb/kube-arangodb/tree/1.2.32) (2023-08-07)
 - (Feature) Backup lifetime - remove Backup once its lifetime has been reached